requirements.txt
torch
git+https://github.com/nielsrogge/transformers.git@add_dpt_redesign#egg=transformers
numpy
Pillow
gradio
blurgenerator